separated components in Section files, removed unused files
