+++ src/components/Footer/styles.ts
@@ -9,12 +9,12 @@ export const FooterContainer = styled.div`
 `
 
 export const Content = styled.div`
+  ${mixins.contentSpace};
   display: grid;
   grid-template-columns: 100%;
   grid-template-rows: auto auto;
   max-width: 80vw;
   margin: 0 auto;
-  ${mixins.contentSpace};
   padding: 0 1rem;
 `
 export const ContentRow = styled.div`
