--- src/components/shared/Neumorphic.tsx
@@ -1,9 +1,0 @@
-import styled from 'styled-components'
-
-interface BaseProps {
-  radius?: any
-}
-
-const Bordered = styled.div<BaseProps>`
-  border-radius: ${props => props.radius || '51px'};
-`
