+++ src/components/shared/Typography.ts
@@ -2,7 +2,6 @@ import styled from 'styled-components'
 import { fonts } from '@theme/styles'
 
 export const Paragraph = styled.p`
-  /* color: ${props => props.theme.textColor}; */
   font-family: ${fonts.muli};
   font-size: 1.1rem;
   letter-spacing: 0.03rem;
@@ -16,7 +15,6 @@ export const Title = styled.p`
   text-transform: uppercase;
   letter-spacing: 2px;
   font-weight: 300;
-  /* color: ${props => props.theme.textColor}; */
 `
 export const TextContent = styled.div`
   max-width: 50rem;
