+++ src/sections/About/index.ts
@@ -1,0 +1,3 @@
+import About from './About'
+
+export default About
