+++ src/sections/Archive/index.ts
@@ -1,0 +1,3 @@
+import Archive from './Archive'
+
+export default Archive
