+++ src/sections/Featured/index.ts
@@ -1,0 +1,3 @@
+import Featured from './Featured'
+
+export default Featured
