+++ src/sections/Resume/index.ts
@@ -1,0 +1,3 @@
+import Resume from './Resume'
+
+export default Resume
